// src/pdfs/ny/ucsnc1-name-change-and-or-sex-designation-change-petition-for-individual-adult/schema.ts
/** Auto-generated from ucsnc1-name-change-and-or-sex-designation-change-petition-for-individual-adult.pdf — do not edit */
import {
  PDFButton,
  PDFCheckBox,
  PDFDropdown,
  PDFRadioGroup,
  PDFTextField,
} from "@cantoo/pdf-lib";

export const pdfSchema = {
  Reset_Form: PDFButton,
  SpousalSupportPayments: PDFRadioGroup,
  PreviousNameChangePetition: PDFRadioGroup,
  SealingRequest: PDFRadioGroup,
  "SealingRequest-specify": PDFTextField,
  SupportingDocument: PDFRadioGroup,
  Print_Form: PDFButton,
  courtType: PDFDropdown,
  courtCounty: PDFDropdown,
  indexNumber: PDFTextField,
  shouldChangeName: PDFCheckBox,
  shouldChangeSexDesignation: PDFCheckBox,
  newName: PDFTextField,
  birthplace: PDFTextField,
  hasConviction: PDFRadioGroup,
  courtOfConviction: PDFTextField,
  crime: PDFTextField,
  hasBankruptcy: PDFRadioGroup,
  hasJudgementsOrLiens: PDFRadioGroup,
  partyToAction: PDFRadioGroup,
  partyToActionDetails: PDFTextField,
  currentlyMarried: PDFRadioGroup,
  previouslyMarried: PDFRadioGroup,
  childrenUnder21: PDFRadioGroup,
  childSupport: PDFRadioGroup,
  childSupportPayments: PDFRadioGroup,
  childSupportOwed: PDFTextField,
  courtIssuingChildSupportOrder: PDFTextField,
  childSupportCollectionsUnit: PDFTextField,
  spousalSupport: PDFRadioGroup,
  spousalSupportOwed: PDFTextField,
  courtIssuingSpousalSupportOrder: PDFTextField,
  hasPreviousNameChange: PDFRadioGroup,
  previousNameReason: PDFTextField,
  reasonForChangingName: PDFTextField,
  newSex: PDFDropdown,
  previousSexDesignationChangePetition: PDFRadioGroup,
  previousSexDesignationChangePetitionReason: PDFTextField,
  sexChangeDesignationReason: PDFTextField,
  providesSexChangeDesignationReason: PDFRadioGroup,
  age: PDFTextField,
  dateOfBirth: PDFTextField,
  residenceAddress: PDFTextField,
  supportingDocumentsDescription: PDFTextField,
  signatureDate: PDFTextField,
  signatureDay: PDFDropdown,
  signatureMonth: PDFDropdown,
  signatureYearEnding: PDFDropdown,
  oldName: PDFTextField,
} as const;

export type PdfFieldName = keyof typeof pdfSchema;